requirements.txt
spotipy
tidalapi
tqdm
python-dotenv
transliterate